requirements.txt
Django==2.1.4
gunicorn==19.9.0
whitenoise==4.1.2
djangorestframework==3.9.0
markdown==3.0.1
django-filter==2.0.0
factory-boy==2.11.1